+++ pkg/verifier/publickey/publickey.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package certificate
+package publickey
 
 import (
 	"crypto"
@@ -23,47 +23,53 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/sigstore/rekor-tiles/pkg/types/verifier/identity"
+	"github.com/sigstore/rekor-tiles/pkg/verifier/identity"
 	"github.com/sigstore/sigstore/pkg/cryptoutils"
 )
 
-// Certificate implements verifier.Verifier
-type Certificate struct {
-	cert *x509.Certificate
+// PublicKey implements verifier.Verifier
+type PublicKey struct {
+	key crypto.PublicKey
 }
 
-func NewVerifier(r io.Reader) (*Certificate, error) {
+func NewVerifier(r io.Reader) (*PublicKey, error) {
 	if r == nil {
-		return nil, errors.New("certificate reader is nil")
+		return nil, errors.New("public key reader is nil")
 	}
 	derVerifier, err := io.ReadAll(r)
 	if err != nil {
 		return nil, err
 	}
-	cert, err := x509.ParseCertificate(derVerifier)
+
+	key, err := x509.ParsePKIXPublicKey(derVerifier)
 	if err != nil {
-		return nil, fmt.Errorf("parsing certificate: %v", err)
+		return nil, fmt.Errorf("parsing public key: %v", err)
 	}
-	return &Certificate{cert: cert}, nil
+	return &PublicKey{key: key}, nil
 }
 
-func (c Certificate) String() string {
-	encoded, err := cryptoutils.MarshalCertificateToPEM(c.cert)
+func (k PublicKey) String() string {
+	encoded, err := cryptoutils.MarshalPublicKeyToPEM(k.key)
 	if err != nil {
 		return ""
 	}
 	return string(encoded)
+
 }
 
-func (c Certificate) PublicKey() crypto.PublicKey {
-	return c.cert.PublicKey
+func (k PublicKey) PublicKey() crypto.PublicKey {
+	return k.key
 }
 
-func (c Certificate) Identities() ([]identity.Identity, error) {
-	digest := sha256.Sum256(c.cert.Raw)
+func (k PublicKey) Identities() ([]identity.Identity, error) {
+	pkixKey, err := cryptoutils.MarshalPublicKeyToDER(k.key)
+	if err != nil {
+		return nil, err
+	}
+	digest := sha256.Sum256(pkixKey)
 	return []identity.Identity{{
-		Crypto:      c.cert,
-		Raw:         c.cert.Raw,
+		Crypto:      k.key,
+		Raw:         pkixKey,
 		Fingerprint: hex.EncodeToString(digest[:]),
 	}}, nil
 }
